scrach-work/nvidia-nemo-experimenting/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
altair==4.1.0
antlr4-python3-runtime==4.8
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.1
atomicwrites==1.4.0
attrdict==2.0.1
attrs==20.3.0
audioread==2.1.9
Automat==20.2.0
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==19.10b0
bleach==3.3.0
boto3==1.20.49
botocore==1.23.49
braceexpand==0.1.7
cachetools==4.2.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cycler==0.10.0
Cython==0.29.26
datasets==1.16.1
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.17.1
editdistance==0.6.0
entrypoints==0.3
fasttext==0.9.2
filelock==3.0.12
Flask==2.0.2
flatbuffers==2.0
frozendict==2.3.0
frozenlist==1.2.0
fsspec==2021.11.1
ftfy==6.0.3
future==0.18.2
g2p-en==2.1.0
gast==0.4.0
gdown==4.2.1
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
huggingface-hub==0.2.1
hydra-core==1.1.1
hyperlink==21.0.0
idna==2.10
imagesize==1.3.0
importlib-metadata==4.8.2
incremental==17.5.0
inflect==5.4.0
iniconfig==1.1.1
ipadic==1.0.0
ipykernel==5.5.3
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
itsdangerous==2.0.1
jedi==0.18.0
jieba==0.42.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.0
kaldi-io==0.9.4
kaldi-python-io==1.2.2
kaldiio==2.17.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
latexcodec==2.0.1
libclang==12.0.0
librosa==0.8.1
llvmlite==0.38.0
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
matplotlib==3.3.4
mecab-python3==1.0.3
mistune==0.8.4
mpmath==1.2.1
multidict==5.2.0
multiprocess==0.70.12.2
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nemo-toolkit @ git+https://github.com/NVIDIA/NeMo.git@7a8c050973434b96df1016eb355583762b7aff8e
nest-asyncio==1.5.1
networkx==2.5
nltk==3.6.7
notebook==6.3.0
numba==0.55.1
numpy==1.21.2
oauthlib==3.1.1
omegaconf==2.1.1
onnx==1.10.2
OpenCC==1.1.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
pangu==4.0.6.1
parameterized==0.8.1
parso==0.8.1
pathspec==0.9.0
pathtools==0.1.2
pesq==0.0.3
pickleshare==0.7.5
Pillow==8.1.2
pip-api==0.0.27
pipreqs==0.4.11
pluggy==0.13.1
pooch==1.6.0
portalocker==2.3.2
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.17
protobuf==3.19.1
psutil==5.9.0
py==1.10.0
pyannote.core==4.3
pyannote.database==4.1.1
pyannote.metrics==3.2
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.1
pybtex==0.24.0
pybtex-docutils==1.0.1
pycparser==2.20
pycryptodome==3.9.9
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.8.1
PyHamcrest==2.0.2
pyparsing==2.4.7
pypinyin==0.45.0
pyrsistent==0.17.3
PySocks==1.7.1
pystoi==0.3.3
pytest==6.2.3
pytest-runner==5.3.1
python-dateutil==2.8.1
pytorch-lightning==1.5.9
pytz==2021.1
pyvis==0.1.9
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
rapidfuzz==1.9.1
regex==2022.1.18
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.5.1
sacrebleu==2.0.0
sacremoses==0.0.45
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
sentencepiece==0.1.96
sentry-sdk==1.5.4
shellingham==1.4.0
shortuuid==1.0.8
simplejson==3.17.6
six==1.15.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
sox==1.4.1
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.4.1
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
syllables==1.0.3
sympy==1.9
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
toolz==0.11.1
torch==1.8.1+cu102
torch-stft==0.1.4
torchaudio==0.8.1
torchmetrics==0.7.1
torchvision==0.9.1+cu102
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
transformers==4.5.1
typed-ast==1.5.2
typer==0.4.0
typing-extensions==3.10.0.0
Unidecode==1.3.2
urllib3==1.26.2
wandb==0.12.10
wcwidth==0.2.5
webdataset==0.1.62
webencodings==0.5.1
Werkzeug==2.0.2
wget==3.2
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wrapt==1.13.3
xxhash==2.0.2
yarg==0.1.9
yarl==1.7.2
yaspin==2.1.0
youtokentome==1.0.6
zipp==3.6.0
zope.interface==5.2.0
